+++ flask_restalchemy/serialization/modelserializer.py
@@ -1,21 +1,11 @@
-from typing import Union
-
-from sqlalchemy.orm.dynamic import AppenderMixin
-from sqlalchemy.sql import sqltypes
-
-from flask_restalchemy.serialization.serializer import Serializer, DateTimeSerializer, is_datetime_field, is_enum_field, \
-    EnumSerializer
+from .fields import Field
+from .serializer import Serializer
 
 
 class ModelSerializer(Serializer):
     """
     Serializer for SQLAlchemy Declarative classes
     """
-
-    DEFAULT_SERIALIZERS = [
-        (is_datetime_field, DateTimeSerializer),
-        (is_enum_field, EnumSerializer)
-    ]
 
     def __init__(self, model_class):
         """
@@ -28,30 +18,33 @@ class ModelSerializer(Serializer):
             field = self._fields.setdefault(column, Field())
             # Set a serializer for fields that can not be serialized by default
             if field.serializer is None:
-                serializer = self._get_default_serializer(self.model_columns.get(column))
-                if serializer:
-                    field._serializer = serializer
-
-    def get_model_name(self) -> str:
-        return self._mapper_class.__name__
+                from flask_restalchemy import Api
+                serializer = Api.find_column_serializer(self.model_columns.get(column))
+                field._serializer = serializer
 
     @property
-    def model_columns(self) -> dict:
+    def model_class(self):
+        return self._mapper_class
+
+    @property
+    def model_columns(self):
         return self._mapper_class.__mapper__.c
 
     def dump(self, model):
+        """
+        Create a serialized dict from a Declarative model
+
+        :param DeclarativeMeta model: the model to be serialized
+
+        :rtype: dict
+        """
         serial = {}
         for attr, field in self._fields.items():
-            if field is None:
-                serial[attr] = getattr(model, attr)
-                continue
             if field.load_only:
                 continue
             value = getattr(model, attr) if hasattr(model, attr) else None
-            if value is None:
-                serialized = None
-            elif field.serializer:
-                serialized = field.serializer.dump(value)
+            if field:
+                serialized = field.dump(value)
             else:
                 serialized = value
             serial[attr] = serialized
@@ -69,31 +62,18 @@ class ModelSerializer(Serializer):
         """
         if existing_model:
             model = existing_model
-            # If deserialization must update an existing model, check if primary key is the same
-            pk_name = self._get_pk_name(existing_model)
-            assert getattr(existing_model, pk_name) == serialized[pk_name], \
-                "Primary key value of serialized nested object is inconsistent"
         else:
             model = self._mapper_class()
         for field_name, value in serialized.items():
             field = self._fields[field_name]
             if field.dump_only:
                 continue
-            if value is None:
-                deserial_value = value
-            elif field.serializer:
-                if isinstance(field.serializer, ModelSerializer):
-                    if existing_model:
-                        existing_nested = getattr(existing_model, field_name)
-                    else:
-                        existing_nested = self._find_existing_model(field, value)
-                    deserial_value = field.serializer.load(value, existing_nested)
-                else:
-                    deserial_value = field.serializer.load(value)
-            else:
-                deserial_value = value
-            setattr(model, field_name, deserial_value)
+            deserialized = field.load(value)
+            setattr(model, field_name, deserialized)
         return model
+
+    def get_model_name(self) -> str:
+        return self._mapper_class.__name__
 
     def before_put_commit(self, model, session):
         pass
@@ -106,11 +86,6 @@ class ModelSerializer(Serializer):
 
     def after_post_commit(self, model, session):
         pass
-
-    def _get_default_serializer(self, column):
-        for check_type, serializer_class in self.DEFAULT_SERIALIZERS:
-            if check_type(column):
-                return serializer_class(column)
 
     @classmethod
     def _get_declared_fields(cls) -> dict:
@@ -125,128 +100,3 @@ class ModelSerializer(Serializer):
             if isinstance(value, Field):
                 fields[attr_name] = value
         return fields
-
-    def _find_existing_model(self, field, value):
-        class_mapper = field.serializer._mapper_class
-        pk_name = self._get_pk_name(class_mapper)
-        pk = value.get(pk_name)
-        if pk:
-            return class_mapper.query.get(pk)
-        else:
-            return None
-
-    def _get_pk_name(self, existing_model):
-        primary_keys = existing_model.__mapper__.primary_key
-        assert len(primary_keys) == 1, "Nested object must have exactly one primary key"
-        pk_name = primary_keys[0].key
-        return pk_name
-
-
-class Field(object):
-    """
-    Configure a ModelSerializer field
-    """
-
-    def __init__(self, dump_only=False, load_only=False, serializer=None):
-        self.dump_only = dump_only
-        self.load_only = load_only
-        self._serializer = serializer
-
-    @property
-    def serializer(self):
-        return self._serializer
-
-
-class NestedModelField(Field):
-    """
-    A field to Dump and Update nested models.
-    """
-
-    def __init__(self, declarative_class, **kw):
-        super().__init__(**kw)
-        if self._serializer is None:
-            self._serializer = ModelSerializer(declarative_class)
-
-
-class NestedAttributesField(Field):
-    """
-    A read-only field that dump nested object attributes.
-    """
-
-    class NestedAttributesSerializer(Serializer):
-
-        def __init__(self, attributes, many):
-            self.attributes = attributes
-            self.many = many
-
-        def dump(self, value):
-            if self.many:
-                serialized = [self._dump_item(item) for item in value]
-            else:
-                return self._dump_item(value)
-            return serialized
-
-        def _dump_item(self, item):
-            serialized = {}
-            for attr_name in self.attributes:
-                serialized[attr_name] = getattr(item, attr_name)
-            return serialized
-
-        def load(self, serialized):
-            raise NotImplementedError()
-
-
-    def __init__(self, attributes: Union[tuple, dict], many=False, **kw):
-        super().__init__(serializer=self.NestedAttributesSerializer(attributes, many), **kw)
-        assert 'serializer' not in kw, "NestedAttributesField does not support custom Serializer"
-
-
-class PrimaryKeyField(Field):
-    """
-    Convert relationships in a list of primary keys (for serialization and deserialization).
-    """
-
-    class PrimaryKeySerializer(Serializer):
-
-        def __init__(self, declarative_class):
-            self.declarative_class = declarative_class
-            self._pk_column = get_model_pk(self.declarative_class)
-
-        def load(self, serialized):
-            pk_column = self._pk_column
-            return self.declarative_class.query.filter(pk_column.in_(serialized)).all()
-
-        def dump(self, value):
-            pk_column = self._pk_column
-            if is_tomany_attribute(value):
-                serialized = [getattr(item, pk_column.key) for item in value]
-            else:
-                return getattr(value, pk_column.key)
-            return serialized
-
-    def __init__(self, declarative_class, **kw):
-        super().__init__(serializer=self.PrimaryKeySerializer(declarative_class), **kw)
-
-
-def get_model_pk(declarative_class):
-    """
-    Get the primary key Column object from a Declarative model class
-
-    :param Type[DeclarativeMeta] declarative_class: a Declarative class
-
-    :rtype: Column
-    """
-    primary_keys = declarative_class.__mapper__.primary_key
-    assert len(primary_keys) == 1, "Nested object must have exactly one primary key"
-    return primary_keys[0]
-
-
-def is_tomany_attribute(value):
-    """
-    Check if the Declarative relationship attribute represents a to-many relationship.
-
-    :param value: a SQLAlchemy Declarative class relationship attribute
-
-    :rtype: bool
-    """
-    return isinstance(value, (list, AppenderMixin))